Implemented seat, started converting stations
Implemented the Seat with an Interaction. Stations can have a reference to the local seat component and use it for a Manned Character which is an abstract Get for all Stations. Working on converting Driver station, Seat adds Pre-Req and is a Traversable similar to Tank. Currently a bug with a null character reference.