More bug-fixing, Inventory not working
Fixed positional bugs (was mixing up local & global position in navigation logic).

Discovered a bug with inventories & characters, characters need to add their pre-requisites. This is simple, we'll just send the pre-req added to the state machine by the tank to the interaction manager.

ALMOST ready to pull fk me